cli: separated functions for executing cwl files
